backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
biopython
numpy
rdkit-pypi
python-multipart
requests
